WIP: Created event lookup object, so I can search NOX events by timepoint
Aggregating events into "buckets" makes sense for
trips on links, but doesn't makse sense for NOX measurements at
different times of day.

Instead, I need a way to find the NOX event at the upper-bound of
a particular timepoint; i.e. user slides to 12:00 Noon, and I want to
show the NOX situation reflected by events up to that point in time
but not past it.

After some research it seems like a sorted binary-tree, ordered by event
time, makes the most sense. I can then do a binary-search to find the
closest upper-limit event to a particular time point and use the NOX
measurement from that event.

For each node. We'll see how performant this is. If it's bad we can
move events to a database and query it with a WHERE clause, as long
as events are indexed on their time attribute.